requirements.txt
python-gitlab
slack_sdk
pyyaml
kubernetes
